获取字段信息（带权限）

+++ src/main/java/com/ioc/datasupport/warehouse/domain/DlRescataColumn.java
@@ -71,7 +71,7 @@ public class DlRescataColumn extends Model<DlRescataColumn> {
      * 字段长度
      */
     @TableField("COLUMN_LENGTH")
-    private Long columnLength;
+    private Integer columnLength;
 
     /**
      * 小数位数
